refactor: unify reporting contracts and toolchain docs
[Summary]
统一 `libs/` 中 reporting/query/transport 的整数与签名口径，补齐 Android 侧适配。
同时收拢 `docs/toolchain/` 与相关说明文档，保持 capability owner 与宿主接线口径一致。

[Component Versions]
- android-presentation: v0.2.7
- libs/tracer_core: changed

[Changed & Refactored]
- 统一 reporting/query/transport 的契约与宿主接线说明，减少整数类型与签名口径分叉
- 更新 Android 侧适配与 presentation/history 文档，使 runtime bridge 相关说明对齐新 reporting 合同
- 收拢 toolchain/docs 中与 reporting contract、capability routing 相关的说明入口

[Verification]
- git diff --cached --check

Release-Version: v0.9.3

+++ libs/tracer_core/src/infra/platform/windows/windows_platform_clock.cpp
@@ -1,6 +1,7 @@
 // infrastructure/platform/windows/windows_platform_clock.cpp
 #include "infra/platform/windows/windows_platform_clock.hpp"
 
+#include <cstdint>
 #include <ctime>
 #include <string>
 
@@ -73,7 +74,7 @@ auto WindowsPlatformClock::LocalUtcOffsetMinutes() const -> int {
   }
 
   const auto kOffsetSeconds =
-      static_cast<long long>(kLocalAsEpoch - kUtcAsLocalEpoch);
+      static_cast<std::int64_t>(kLocalAsEpoch - kUtcAsLocalEpoch);
   return static_cast<int>(kOffsetSeconds / kSecondsPerMinute);
 }
